Initial commit
This includes the code extracted from the main app with changes to
remove solidus as a dependency. This also replaces paperclip with
active_storage. It includes specs and the setup needed to boot up a
dummy app while testing.

Two components had to be removed, due to heavy solidus dependencies:
- Taxon
- Products

The rest of the components should be working fine.